client: split sdk_bypass.go
